requirements.txt
-i https://pypi.org/simple
amqp==5.1.1 ; python_version >= '3.6'
async-timeout==4.0.2 ; python_version >= '3.6'
billiard==3.6.4.0
celery==5.2.7
cfgv==3.3.1 ; python_full_version >= '3.6.1'
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-bootstrap-templates==1.0.7
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
deprecated==1.2.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
distlib==0.3.6
filelock==3.8.0 ; python_version >= '3.7'
flask==2.2.2 ; python_version >= '3.7'
gunicorn==20.1.0
identify==2.5.8 ; python_version >= '3.7'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
kombu==5.2.4 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
nodeenv==1.7.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==1.23.4
packaging==21.3 ; python_version >= '3.6'
pandas==1.5.1
patsy==0.5.3
platformdirs==2.5.3 ; python_version >= '3.7'
plotly==5.11.0 ; python_version >= '3.6'
plotly-express==0.4.1
pre-commit==2.20.0
prompt-toolkit==3.0.32 ; python_full_version >= '3.6.2'
psycopg2-binary==2.9.5
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.6
pyyaml==6.0 ; python_version >= '3.6'
redis==4.3.4
rq==1.11.1
scipy==1.9.3 ; python_version >= '3.8'
setuptools==65.5.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.43
statsmodels==0.13.5 ; python_version >= '3.7'
tenacity==8.1.0 ; python_version >= '3.6'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
vine==5.0.0 ; python_version >= '3.6'
virtualenv==20.16.6 ; python_version >= '3.6'
wcwidth==0.2.5
werkzeug==2.2.2 ; python_version >= '3.7'
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requests
dash-mantine-components
pyarrow
fuzzywuzzy
python-Levenshtein
